redid site flow, made pages for each function
